refactor: streamline pulumi app layout and docker stages
